pyproject.toml
[project]
authors = [
    {name = "Colsrch", email = "colsrch@foxmail.com"},
]
requires-python = ">=3.9,<4.0"
license = {text = "AGPL-3"}
dependencies = [
    "graia-ariadne==0.9.4",
    "graia-scheduler==0.0.8",
    "graia-saya==0.0.16",
    "richuru==0.1.1",
    "arclet-alconna==1.2.0.7",
    "arclet-alconna-graia==0.6.2",
    "prompt-toolkit~=3.0.31",
    "prettytable~=3.4.1",
    "retrying~=1.3.3",
    "peewee~=3.15.2",
    "pillow~=9.2.0",
    "jsonpath~=0.82",
    "pip~=22.2.2",
]

[project.optional-dependencies]
mysql = [
    "pymysql~=1.0.2",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[project.urls]
repository = "https://github.com/MadokaProject/Madoka"
documentation = "https://madoka.colsrch.cn"

[tool]

[tool.pdm]

[tool.pdm.scripts]
start.cmd = "python main.py"
start.help = "启动 Madoka"
clear.cmd = "rm -rf **/__pycache__"
clear.help = "清理 __pycache__"

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=2.20.0",
    "black>=22.8.0",
    "flake8>=5.0.4",
    "isort>=5.10.1",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.pyright]
ignore = [
    "docs/**",
    "**/__pypackages__/**/*.py",
    "**/test*/**/*.py",
    "**/tmp/**",
    "**/temp/**/*.py",
    "app/util/network.py",
]

[tool.pydocstyle]
match-dir = "(?!\\.)(?!test).*"
